Propagates color choice env variables to hooks and commands
This change propagats color choice of prek to the hooks and commands
it executes. It does it by setting the environment variables based on
Always/Auto/Never.

For now we are using just pretty "standard" env variables used in a
number of tools:

* NO_COLOR -> obvious (Never)
* FORCE_COLOR -> when you want to force the color regardless of
  detection (used for example by ruff)a (Always)
* CLICOLOR -> when you want to "hint" to other rust binaries that
  the color is available (used by colored crate) (Auto)
* CLICOLOR_FORCE -> when you want to force setting color for coloured
  crate supported binaries (Auto)

While this does not solve ruff colouring when pseudo terminal is not
allocated, it does color the output even if there is no pseudo-terminal.

With this change, for example ruff output will be coloured when
--color always is used (which did not work before) - you had to manually
set FORCE_COLOR=1 before running prek to make it works..

This is not a complete solution yet. The problem currently is that we
are capturing output from the commands that we execute. Even if we
allocate pseudo terminal and pass it as "usual" this will not solve
the problem because pseudo terminal in this case will be only used by
stdin and capturing the output will cause the stdout and stderr to
be pipes (i.e, isatty returns False). What we really want in the final
solution is to create pseudo terminal, pass it (including stdout and
stderr) to the command, and then read the output and error from the
pseudo terminal. This will make it possible for auto detection
of colour by all tools to work properly.

The pty-check prek hook is also added to allow to test easily what
are the terminal variables and terminal properties passed when you
run hook.

Related: #440